requirements.txt
# ─────────────────────────────────────────────
#  Voice Bot – Python Dependencies
# ─────────────────────────────────────────────

# API Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
aiofiles==23.2.1

# ASR – Whisper
openai-whisper==20231117
# ffmpeg-python==0.2.0   # install ffmpeg via system package manager

# NLP / Intent Classification
transformers==4.41.2
torch==2.3.0
datasets==2.19.1
scikit-learn==1.5.0
numpy==1.26.4
pandas==2.2.2

# Text-to-Speech
gTTS==2.5.1
pyttsx3==2.90        # offline TTS fallback

# Audio utilities
pydub==0.25.1
soundfile==0.12.1
librosa==0.10.2
jiwer==3.0.4         # WER calculation

# Visualisation
matplotlib==3.9.0
seaborn==0.13.2

# Utilities
python-dotenv==1.0.1
loguru==0.7.2
httpx==0.27.0
tqdm==4.66.4
Pillow==10.3.0

# Dev / testing
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
